feat: initial commit
This commit adds angular cli project and the base feature that I need to explain my proof of concept with mobx.
All I want is to show how powerful and simple mobx could be.

In this commit I also include the readme of the project with the brief description about the project and about mobx.